FIX Delete relation only if it exists

+++ src/Form.php
@@ -918,7 +918,10 @@ class Form implements Renderable
                         $instance = $relation->findOrNew(Arr::get($related, $keyName));
 
                         if ($related[static::REMOVE_FLAG_NAME] == 1) {
-                            $instance->delete();
+                            if (Arr::get($related, $keyName)) {
+                                // only if exists id
+                                $instance->delete();
+                            }
 
                             continue;
                         }
